Uc 2 maximum float

+++ src/com/bridgelabz/generic/Maximum.java
@@ -6,15 +6,19 @@ public class Maximum {
 
         FindMax max = new FindMax();
         Scanner scan = new Scanner(System.in);
-            System.out.print("Enter First Value: ");
-            Integer first = scan.nextInt();
+	    
+           System.out.print("Enter First Value: ");
+           Float first = scan.nextFloat();
 
-            System.out.print("Enter Second Value: ");
-            Integer second = scan.nextInt();
+           System.out.print("Enter Second Value: ");
+           Float second = scan.nextFloat();
 
-            System.out.print("Enter Third Value: ");
-            Integer third = scan.nextInt();
+           System.out.print("Enter Third Value: ");
+           Float third = scan.nextFloat();
 
-	    max.findMax(first,second,third);
-  }
-}
+            max.findMax(first,second,third);
+        }
+ }
+
+
+
